Add setControlPortPasswd() to RelativeLink/RelativeLink.sh.
This adds documentation on how to set the ``TOR_CONTROL_PASSWD``
environment variable, as well as a bash function,
``setControlPortPasswd()``, which easily sets it.

Using a bash function is necessary in order to safely set
TOR_CONTROL_PASSWD within the function, such that it isn't accidentally
``export``ed into the shell which called the script (where other
programs would be able to steal it). Inside the function, is checked to
see if the odd quoting syntax was done properly, and, if not, complains
and tells the user how to properly set the
``TOR_CONTROL_PASSWD``. Otherwise, the correctly set password is given
before the command to start Firefox, i.e.:

  TOR_CONTROL_PASSWD=${TOR_CONTROL_PASSWD} ./firefox [...]

 * ADDS a new bash function ``setControlPortPasswd()`` to RelativeLink.sh.
 * FIXES #10178 by making it really hard for people to continue DoingItWrong™.
 * FIXES #11751 by adding an easy way to set the TOR_CONTROL_PASSWD, as well
   as documentation.