Bug 21754: Remove old GCC toolchain and macOS SDK
We remove all hints to our old GCC-based toolchain for macOS and switch
our descriptors to use the 10.7 SDK as 10.9 is the minimum requirement
for Firefox 52 ESR now.

We clean up the tor macOS descriptor a bit as well, removing unused
environment variables and replacing '/home/debian' with '$HOME' as done
in the utils descriptor.