Bug 20683: Integrate Selfrando into alpha Linux builds
Selfrando is a new defense against code reuse attacks developed by the
Redactor and Readactor++ people. We should give it a wider testing
audience by including it in the alpha series.

This is currently only available for 64bit Linux builds, though.
Supporting other platforms and architectures is work in progress.