Bug 22361: fix hardening of libraries built in linux/gitian-utils.yml
With the changes to integrate Selfrando (#20683), we are using our build
of gcc and binutils to build other libraries in the linux/gitian-utils.yml
descriptor, which removed the use of the hardening wrappers. We fix that
by adding the hardening wrappers to our builds of gcc and binutils.
Because we add the wrappers to gcc and binutils zip files, we don't need
to add them anymore in the other descriptors where they are used.